# api/openapi.yaml
openapi: 3.1.0
info:
  title: ThinkPixelGR API
  version: 0.2.0
  description: Policy-driven evaluation for LLM-adjacent content.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - {}
paths:
  /health/live:
    get:
      operationId: live
      security: []
      responses:
        "200": { $ref: "#/components/responses/Health" }
  /health/ready:
    get:
      operationId: ready
      security: []
      responses:
        "200": { $ref: "#/components/responses/Health" }
  /v1/policies:
    get:
      operationId: listPolicies
      responses:
        "200":
          description: Loaded canonical policy identifiers.
          content:
            application/json:
              schema:
                type: object
                required: [policies]
                properties:
                  policies: { type: array, items: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/evaluations:
    post:
      operationId: evaluate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluationRequest" }
      responses:
        "200":
          description: A policy decision. Blocks are successful evaluations.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluationResponse" }
        "400":
          description: Invalid request or unknown policy/profile.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "504":
          description: The caller or overall evaluation context expired; no policy failure mode was applied.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: A deployment-issued credential mapped server-side to tenant and API capabilities.
  responses:
    Health:
      description: Healthy.
      content:
        application/json:
          schema:
            type: object
            properties: { status: { type: string, const: ok } }
    Unauthorized:
      description: Bearer authentication is missing or invalid.
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The authenticated principal lacks the requested tenant or API capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
          additionalProperties: false
      additionalProperties: false
    EvaluationRequest:
      type: object
      required: [request_id, stage, content]
      properties:
        request_id: { type: string, minLength: 1 }
        stage:
          type: string
          enum: [pre_request, pre_model, post_model, pre_tool, post_tool, pre_retrieval, post_retrieval, ingestion]
        tenant_id:
          type: string
          description: Must be within the authenticated principal's server-side tenant scope. Omission requires an explicit tenantless scope.
        guardrails:
          type: object
          properties:
            profile: { type: string }
            policies: { type: array, items: { type: string } }
        content: { $ref: "#/components/schemas/Content" }
        subject:
          type: object
          description: Untrusted evaluation input; it does not establish identity or authority.
          additionalProperties: true
        target: { type: object, additionalProperties: true }
        metadata: { type: object, additionalProperties: true }
    Content:
      type: object
      properties:
        type: { type: string }
        text: { type: string }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string }
              content: { type: string }
        data: { type: object, additionalProperties: true }
    EvaluationResponse:
      type: object
      required: [evaluation_id, request_id, decision, applied_policies, findings, timing]
      properties:
        evaluation_id: { type: string }
        request_id: { type: string }
        decision:
          type: object
          required: [action, reason]
          properties:
            action: { type: string, enum: [allow, block, redact, monitor] }
            reason: { type: string }
            transformed_content: { $ref: "#/components/schemas/Content" }
        applied_policies: { type: array, items: { type: string } }
        findings:
          type: array
          items:
            type: object
            required: [detector, category, confidence]
            properties:
              detector: { type: string }
              category: { type: string }
              confidence: { type: number, minimum: 0, maximum: 1 }
              severity: { type: string }
              attributes:
                type: object
                description: Detector evidence metadata. It never contains matched secret values.
                additionalProperties: true
              locations:
                type: array
                items:
                  type: object
                  required: [path, start, end]
                  properties:
                    path: { type: string }
                    start: { type: integer, minimum: 0 }
                    end: { type: integer, minimum: 0 }
        detector_failures:
          type: array
          description: Content-free evidence for detector or policy execution failures.
          items:
            type: object
            required: [policy, scope, code, failure_mode, timeout_ms, retryable]
            properties:
              policy: { type: string, minLength: 1 }
              detector: { type: string, minLength: 1 }
              scope: { type: string, enum: [policy, detector] }
              code:
                type: string
                enum: [POLICY_TIMEOUT, DETECTOR_TIMEOUT, DETECTOR_UNAVAILABLE, DETECTOR_INVALID_RESPONSE, DETECTOR_UNSUPPORTED_INPUT, DETECTOR_INTERNAL_ERROR]
              failure_mode: { type: string, enum: [open, closed, monitor] }
              timeout_ms: { type: integer, minimum: 1 }
              retryable: { type: boolean }
            additionalProperties: false
        timing:
          type: object
          required: [total_ms, detectors]
          properties:
            total_ms: { type: integer, minimum: 0 }
            detectors:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
            policies:
              type: object
              additionalProperties: { type: integer, minimum: 0 }